Refactor integration tests: consolidate utility functions and classes into `utils.py`, streamline imports, and enhance template loading functionality.
